fix: type for rewards and counts

+++ greedy.go
@@ -87,8 +87,8 @@ func NewEpsilonGreedy(nArms int, epsilonDecay float64) *EpsilonGreedy {
 	counts := cmap.New()
 
 	for i := 0; i < nArms; i++ {
-		rewards.Set(fmt.Sprint(i), 0)
-		counts.Set(fmt.Sprint(i), 0)
+		rewards.Set(fmt.Sprint(i), float64(0))
+		counts.Set(fmt.Sprint(i), int64(0))
 	}
 	return &EpsilonGreedy{
 		N:       nArms,
